RoadBlock
This might be the end of this route of blog backend. It appears impossible currently to access the file system using rust. I might just hardcode what I expect the code to generate, it should still be possible. Its just frustrating to go from having an automatically extensible system with the individual posts separated by the file system to having to put everything in the same folder and hardcode the categories/etc.

I'm about 7 hours of work behind where I wanted to be right now. Looking bleak.